+++ internal/syntax/ast/returning.go
@@ -1,16 +1,16 @@
 package ast
 
 import (
-	projection1 "github.com/efritz/gostgres/internal/execution/projection"
+	projectionHelpers "github.com/efritz/gostgres/internal/execution/projection"
 	"github.com/efritz/gostgres/internal/execution/queries"
-	"github.com/efritz/gostgres/internal/execution/queries/projection"
+	projection "github.com/efritz/gostgres/internal/execution/queries/projection"
 	"github.com/efritz/gostgres/internal/shared/impls"
 )
 
-func wrapReturning(node queries.Node, table impls.Table, alias string, expressions []projection1.ProjectionExpression) (queries.Node, error) {
-	var aliasedTables []projection1.AliasedTable
+func wrapReturning(node queries.Node, table impls.Table, alias string, expressions []projectionHelpers.ProjectionExpression) (queries.Node, error) {
+	var aliasedTables []projectionHelpers.AliasedTable
 	if alias != "" {
-		aliasedTables = append(aliasedTables, projection1.AliasedTable{
+		aliasedTables = append(aliasedTables, projectionHelpers.AliasedTable{
 			TableName: table.Name(),
 			Alias:     alias,
 		})
