+++ internal/syntax/ast/select.go
@@ -2,14 +2,14 @@ package ast
 
 import (
 	"github.com/efritz/gostgres/internal/execution/expressions"
-	projection1 "github.com/efritz/gostgres/internal/execution/projection"
+	projectionHelpers "github.com/efritz/gostgres/internal/execution/projection"
 	"github.com/efritz/gostgres/internal/execution/queries"
 	"github.com/efritz/gostgres/internal/execution/queries/aggregate"
 	"github.com/efritz/gostgres/internal/execution/queries/combination"
 	"github.com/efritz/gostgres/internal/execution/queries/filter"
 	"github.com/efritz/gostgres/internal/execution/queries/limit"
 	"github.com/efritz/gostgres/internal/execution/queries/order"
-	"github.com/efritz/gostgres/internal/execution/queries/projection"
+	projection "github.com/efritz/gostgres/internal/execution/queries/projection"
 	"github.com/efritz/gostgres/internal/shared/fields"
 	"github.com/efritz/gostgres/internal/shared/impls"
 	"github.com/efritz/gostgres/internal/shared/types"
@@ -24,7 +24,7 @@ type SelectBuilder struct {
 }
 
 type SimpleSelectDescription struct {
-	SelectExpressions []projection1.ProjectionExpression
+	SelectExpressions []projectionHelpers.ProjectionExpression
 	From              *TableExpression
 	Where             impls.Expression
 	Groupings         []impls.Expression
@@ -66,7 +66,7 @@ func (b SelectBuilder) TableExpression() (queries.Node, error) {
 	}
 
 	if b.Select.Groupings != nil {
-		projectedExpressions, err := projection1.ExpandProjection(node.Fields(), b.Select.SelectExpressions)
+		projectedExpressions, err := projectionHelpers.ExpandProjection(node.Fields(), b.Select.SelectExpressions)
 		if err != nil {
 			return nil, err
 		}
