+++ internal/syntax/ast/tables.go
@@ -4,12 +4,12 @@ import (
 	"fmt"
 
 	"github.com/efritz/gostgres/internal/execution/expressions"
-	projection1 "github.com/efritz/gostgres/internal/execution/projection"
+	projectionHelpers "github.com/efritz/gostgres/internal/execution/projection"
 	"github.com/efritz/gostgres/internal/execution/queries"
 	"github.com/efritz/gostgres/internal/execution/queries/access"
 	"github.com/efritz/gostgres/internal/execution/queries/alias"
 	"github.com/efritz/gostgres/internal/execution/queries/joins"
-	"github.com/efritz/gostgres/internal/execution/queries/projection"
+	projection "github.com/efritz/gostgres/internal/execution/queries/projection"
 	"github.com/efritz/gostgres/internal/shared/fields"
 	"github.com/efritz/gostgres/internal/shared/impls"
 )
@@ -110,9 +110,9 @@ func (e *TableExpression) TableExpression() (queries.Node, error) {
 				return nil, fmt.Errorf("wrong number of fields in alias")
 			}
 
-			projectionExpressions := make([]projection1.ProjectionExpression, 0, len(fields))
+			projectionExpressions := make([]projectionHelpers.ProjectionExpression, 0, len(fields))
 			for i, field := range fields {
-				projectionExpressions = append(projectionExpressions, projection1.NewAliasedExpression(expressions.NewNamed(field), columnNames[i]))
+				projectionExpressions = append(projectionExpressions, projectionHelpers.NewAliasedExpression(expressions.NewNamed(field), columnNames[i]))
 			}
 
 			node, err = projection.NewProjection(node, projectionExpressions)
